+++ src/01-one/compute/_lib/isLeap.js
@@ -1,9 +1,9 @@
 //https://www.timeanddate.com/date/leapyear.html
 const isLeapYear = function (year) {
   // no leap years in bc
-  if (year < 0) {
-    return false
-  }
+  // if (year < 0) {
+  //   return false
+  // }
   return year % 4 === 0 && (year % 100 !== 0 || year % 400 === 0)
 };
 export default isLeapYear