+++ src/01-one/parse/text.js
@@ -4,9 +4,11 @@ const parseText = function (txt, tz) {
   let cal = {}
   // normalize it a bit first
   txt = txt.toLowerCase()
+  txt = txt.replace(/([0-9])(th|rd|st|nd)\b/, '$1')
+  txt = txt.replace(/\b(mon|tues?|wed|wednes|thur?s?|fri|sat|satur|sun)(day)?\b/i, '')
   txt = txt.replace(/,/g, '')
+  txt = txt.replace(/ +/g, ' ').trim()
   txt = txt.trim()
-
   for (let i = 0; i < formats.length; i += 1) {
     let m = txt.match(formats[i].reg)
     if (m !== null) {